modify: move `assets.py` into `dev` folder
