+++ apps/draw/draw_funcs/abyss.py
@@ -5,7 +5,7 @@ import discord
 import genshin
 from PIL import Image, ImageDraw
 
-import asset
+import dev.asset as asset
 from apps.draw.draw_funcs import leaderboard
 from apps.draw.utility import (
     draw_dynamic_background,
