+++ apps/draw/draw_funcs/banners.py
@@ -4,7 +4,7 @@ from typing import List
 from discord import Locale
 from PIL import Image, ImageDraw
 
-import asset
+import dev.asset as asset
 from apps.draw.utility import get_cache, get_font
 from apps.text_map import text_map
