+++ apps/draw/draw_funcs/characters.py
@@ -4,7 +4,7 @@ from typing import Dict, List
 import genshin
 from PIL import Image, ImageDraw
 
-import asset
+import dev.asset as asset
 from apps.draw.utility import draw_dynamic_background, get_cache, get_font
 from dev.models import DynamicBackgroundInput
