+++ apps/draw/draw_funcs/check.py
@@ -4,7 +4,7 @@ import discord
 import genshin
 from PIL import Image, ImageDraw
 
-import asset
+import dev.asset as asset
 from apps.draw.utility import circular_crop, get_cache, get_font
 from apps.text_map import text_map
