+++ apps/draw/draw_funcs/diary.py
@@ -5,7 +5,7 @@ import discord
 import genshin
 from PIL import Image, ImageDraw
 
-import asset
+import dev.asset as asset
 from apps.draw.utility import get_font, human_format
 from apps.genshin.utility import convert_ar_to_wl, convert_wl_to_mora
 from apps.text_map import get_month_name, text_map
