+++ apps/draw/draw_funcs/lineup.py
@@ -5,7 +5,7 @@ from discord import Locale
 from genshin.models import LineupPreview
 from PIL import Image, ImageDraw
 
-import asset
+import dev.asset as asset
 from apps.draw.utility import get_cache, get_font
 from apps.text_map import text_map
