+++ apps/draw/utility.py
@@ -10,7 +10,7 @@ import enkanetwork as enka
 from fontTools.ttLib import TTFont
 from PIL import Image, ImageChops, ImageDraw, ImageFont
 
-import asset
+import dev.asset as asset
 from apps.text_map import text_map
 from data.draw.fonts import FONTS
 from dev.models import DefaultEmbed, DynamicBackgroundInput
