+++ ui/genshin/Leaderboard.py
@@ -3,7 +3,7 @@ from typing import Dict, List
 import discord
 from discord import ui, utils
 
-import asset
+import dev.asset as asset
 import config
 import dev.models as models
 from ambr import AmbrTopAPI, Character
