+++ ui/others/settings/Notif.py
@@ -1,7 +1,7 @@
 import discord
 from discord import ui
 
-import asset
+import dev.asset as asset
 import config
 from apps.db import get_user_notif
 from apps.text_map import text_map
